% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_data.R
\docType{data}
\name{covid_data}
\alias{covid_data}
\title{World Covid-19 Data}
\format{
An object of class \code{data.frame} with 47749 rows and 41 columns.
}
\source{
The data below was obtained from https://ourworldindata.org/coronavirus

#' @format A data frame with 47,749 observations and 41 variables:
\itemize{
\item \strong{iso_code}: Country id
\item \strong{continent}: Name of Continent
\item \strong{location}: Country's name
\item \strong{date}: Date of observation recorded
\item \strong{total_cases}: Accumulated total confirmed cases of COVID-19
\item \strong{new_cases}: Daily new positive confirmed cases of COVID-19
\item \strong{new_cases_smoothed}: Daily new confirmed cases of COVID-19 smoothed for a period of 7 days
\item \strong{total_deaths}: Total deaths attributed to COVID-19
\item \strong{new_deaths}: New deaths attributed to COVID-19
\item \strong{new_deaths_smoothed}: New deaths attributed to COVID-19 smoothed for a period of 7 days
\item \strong{total_cases_per_million}: Daily positive COVID-19 new cases recorded
\item \strong{new_cases_per_million}:	New confirmed cases of COVID-19 per 1,000,000 people
\item \strong{new_cases_smoothed_per_million}:	New confirmed cases of COVID-19 smoothed for a period of 7 days per 1,000,000 people
\item \strong{total_deaths_per_million}:	Total deaths recorded to COVID-19 per 1,000,000 people
\item \strong{new_deaths_per_million}:	New deaths recorded to COVID-19 per 1,000,000 people
\item \strong{new_deaths_smoothed_per_million}:	New deaths attributed to COVID-19 smoothed for a period of 7 days per 1,000,000 people
\item \strong{total_tests}:	Total tests for COVID-19	locally
\item \strong{new_tests}:	New tests for COVID-19	locally
\item \strong{new_tests_smoothed}:	New tests for COVID-19 smoothed for a period of 7 days. For countries that don't report testing data on a daily basis, we assume that testing changed equally on a daily basis over any periods in which no data was reported. This produces a complete series of daily figures, which is then averaged over a rolling 7-day window	National government reports
\item \strong{total_tests_per_thousand}:	Total tests for COVID-19 per 1,000 people
\item \strong{new_tests_per_thousand}:	New tests for COVID-19 per 1,000 people
\item \strong{new_tests_smoothed_per_thousand}:	New tests for COVID-19 smoothed for a period of 7 days per 1,000 people
\item \strong{tests_per_case}:	Tests conducted per new confirmed case of COVID-19
\item \strong{positive_rate}:	COVID-19 tested positive
\item \strong{tests_units}:	Units used for testing
\item \strong{stringency_index}:	Government Response Stringency Index
\item \strong{population}:	Population of country
\item \strong{population_density}:	Number of people divided by land area per sqkm
\item \strong{median_age}:	Median age of the population
\item \strong{aged_65_older}:	percentage of populatioon 65 years and older
\item \strong{aged_70_older}:	percentage of populatioon 70 years and older
\item \strong{gdp_per_capita}:	Gross domestic product at purchasing power parity
\item \strong{extreme_poverty}:	Percentage of population living in extreme poverty
\item \strong{cardiovasc_death_rate}:	Death rate from cardiovascular disease
\item \strong{diabetes_prevalence}:	percentage of diabetic patients
\item \strong{female_smokers}:	female smokers ratio
\item \strong{male_smokers}:	male smokers ratio
\item \strong{handwashing_facilities}:	hand washing facilities availabe
\item \strong{hospital_beds_per_thousand}:	Hospital beds per 1,000
\item \strong{life_expectancy}:	Life expectancy in a country
\item \strong{human_development_index}:	human development and standard of living
}
}
\usage{
covid_data
}
\description{
the data is used in the app and is extensive, it not only focuses on the status of corona virus cases but also gives data on other factors that might be contributing to the pandemic.
}
\details{
This dataset countains covid-19 data on \strong{all} countries.
}
\keyword{datasets}
