% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_reactive.R
\name{interact}
\alias{interact}
\title{interact is a function which allows the user to manipulate the bubble chart with the desired variables.}
\usage{
interact(inputId, label, data, selected)
}
\arguments{
\item{inputId}{is the selected ID decided by the user so that it makes easier in plottling data especially in ggplots.}

\item{label, }{to label the data set as deemed appropriate my the user for the purpose intended.}

\item{data, }{the data set being used or required to make the data set.}

\item{selected}{to choose a desired variable before the user has a chance to manipulate the plot}
}
\description{
The interact function helps the users to choose variables from the data set.
}
