% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launches the shiny app from this function}
\usage{
launch_app()
}
\value{
shiny application object
}
\description{
The launch_app() function searches the directory provided to access the app.R from the inst folder
}
\details{
The path has been stored in the appDir (app directory)
}
