% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringency.R
\name{stringency}
\alias{stringency}
\title{stringency function creates a plotly plot by calling a part of the app, the plot shows the stringency index of asian countries and plotly allows the user to engage in a variety of ways.}
\usage{
stringency(input)
}
\arguments{
\item{input}{the given input to draw the plot from the app}
}
\description{
The stringency function allows the users to call the plot and use it, as deemed appropriate.
}
