% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_data.R
\docType{data}
\name{covid_data}
\alias{covid_data}
\title{World Covid-19 Data}
\format{
An object of class \code{data.frame} with 47749 rows and 41 columns.
}
\source{
The data below was obtained from https://ourworldindata.org/coronavirus
}
\usage{
covid_data
}
\description{
the data is used in the app and is extensive, it not only focuses on the status of corona virus cases but also gives data on other factors that might be contributing to the pandemic.
}
\details{
This dataset countains covid-19 data on \strong{all} countries.
}
\keyword{datasets}
