% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidham-package.R
\docType{package}
\name{covidham-package}
\alias{covidham}
\alias{covidham-package}
\title{covidham: Covid-19 Data on Asian Countries}
\description{
(This package contains a shiny application that revolves around different factors revolving around Covid-19. The data only encompuses Asia. Users can use this application to make their own interpretations on the subject.)
}
\author{
\strong{Maintainer}: Muhammad Soban Qasim \email{mqas0003@student.monash.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
